Refactoring
- Move match logic to _wcmatch
- Move all "glob" logic to glob
